fix: add search-utxo command
